+++ Tests.py
@@ -1,12 +1,6 @@
 import unittest
 
-from ally.NewOrder import (
-	TestOrderManual,
-	TestInstrumentConstruction,
-	TestPricingConstruction,
-	TestOrderConstruction
-)
-
+from ally.tests import *
 
 if __name__ == '__main__':
 	unittest.main()
