+++ ally/Ally.py
@@ -4,7 +4,6 @@ from .Auth		import Auth
 from .exception	import ApiKeyException
 from .Api		import setTimeout
 from .Watchlist	import Watchlist
-from .			import NewOrder
