+++ ally/Order/Outstanding.py
@@ -1,5 +1,4 @@
 from ..Api		import AccountEndpoint, RequestType
-from ..FIXML	import FIXML
 
 
 
@@ -23,9 +22,9 @@ class OutstandingOrders ( AccountEndpoint ):
 		if not isinstance(raworders, list):
 			raworders = [raworders]
 
-		orders = [ FIXML ( x ) for x in raworders]
+		# orders = [ FIXML ( x ) for x in raworders]
 
-		return orders
+		return raworders
