+++ ally/Order/__init__.py
@@ -1,5 +1,3 @@
-from .				import order
-from .Submit		import submit
-# from .order			import Order
 from .Outstanding	import orders
-from .Directives	import *
+from .Submit		import submit
+from .order			import Order
