+++ ally/tests.py
@@ -1,0 +1,1 @@
+from .Order.tests import *
